% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium.R
\name{get_driver}
\alias{get_driver}
\title{get_driver}
\usage{
get_driver(port, cache = NULL, ...)
}
\arguments{
\item{port}{specify port number}

\item{cache}{specify cache which will save cookies}

\item{...}{pass aditional arguments to remoteDriver}
}
\description{
get_driver
}
