% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium.R
\name{get_spec_timeline}
\alias{get_spec_timeline}
\title{Get Tweets}
\usage{
get_spec_timeline(driver, screen_name, from, until, end_date, path,
  n_scroll = 5)
}
\arguments{
\item{driver}{Selenium driver}

\item{screen_name}{Give screen_name of user you want to get tweets from}

\item{from}{Give start date as character in this format \code{YYYY-MM-DD}}

\item{until}{Give end date as character in this format \code{YYYY-MM-DD}}

\item{n_scroll}{number of scrolls (Note: if the number of tweets you want to collect is high you may want to increase this)}
}
\description{
Top level function to get and parse tweets
}
