% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_extract_x}
\alias{str_extract_x}
\title{Extract Custom Atrribute from HTML}
\usage{
str_extract_x(x, attr)
}
\arguments{
\item{x}{HTML}

\item{attr}{specify the attribute}
}
\description{
Extract a custom Attribute from HTML
}
